handled cart and checkout for the logged in  users

+++ carts/views.py
@@ -4,6 +4,8 @@ from .models import Cart, CartItem
 
 from store.models import Product, Variation
 from django.core.exceptions import ObjectDoesNotExist
+
+from django.contrib.auth.decorators import login_required
 # Create your views here.
 
 
@@ -13,11 +15,18 @@ def _cart_id(request):
         cart = request.session.create()
     return cart
 
+
+def itemQuantity(request, product_id):
+    pass
+
+
 ############################ Plus button functionality of cart and "Add to cart" button of the product detail page #############################
 
 
 def add_cart(request, product_id):
+    current_user = request.user
     product = Product.objects.get(id=product_id)
+
     product_variation = []
     if request.method == 'POST':
         for item in request.POST:
@@ -28,48 +37,110 @@ def add_cart(request, product_id):
                 # The iexact lookup is used to get records with a specified value. The iexact lookup is case insensitive.
                 variation = Variation.objects.get(
                     product=product, variation_category__iexact=key, variation_value__iexact=value)
+
                 product_variation.append(variation)
+
             except:
                 pass
-    try:
-        cart = Cart.objects.get(cart_id=_cart_id(request))
-    except Cart.DoesNotExist:
-        cart = Cart.objects.create(cart_id=_cart_id(request))
-        cart.save()
 
-    is_cart_item_exits = CartItem.objects.filter(
-        product=product, cart=cart).exists()
-    ex_var_list = []
-    id = []
-    if is_cart_item_exits:
-        cart_item = CartItem.objects.filter(product=product, cart=cart)
-        # existing_variations -> database
-        # current_variation -> product_variation
-        # item_id -> database
+    # if the user is authenticated
+    if current_user.is_authenticated:
 
-        for item in cart_item:
-            existing_variation = item.variations.all()
-            ex_var_list.append(list(existing_variation))
-            id.append(item.id)
+        is_cartItem_exits_for_given_user = CartItem.objects.filter(
+            user=current_user).exists()
 
-        if product_variation in ex_var_list:
-            index = ex_var_list.index(product_variation)
-            item_id = id[index]
-            item = CartItem.objects.get(product=product, id=item_id)
-            item.quantity += 1
-            item.save()
+        if is_cartItem_exits_for_given_user:
+            cart = CartItem.objects.filter(user=current_user)[0].cart
+        else:
+            cart = Cart.objects.create(cart_id=_cart_id(request))
+            cart.save()
+
+        is_given_cart_item_exits = CartItem.objects.filter(
+            product=product, user=current_user).exists()
+        ex_var_list = []
+        id = []
+
+        if is_given_cart_item_exits:
+
+            cart_item = CartItem.objects.filter(
+                product=product, user=current_user, cart=cart)
+
+            for item in cart_item:
+                existing_variation = item.variations.all()
+                ex_var_list.append(list(existing_variation))
+                id.append(item.id)
+
+            if product_variation in ex_var_list:
+                # increase the cart item quantity
+                index = ex_var_list.index(product_variation)
+                item_id = id[index]
+                item = CartItem.objects.get(id=item_id)
+                item.quantity += 1
+                # item.save()
+            else:
+                item = CartItem.objects.create(
+                    product=product, quantity=1, user=current_user, cart=cart)
+                if len(product_variation) > 0:
+                    # item.variations.clear()
+                    item.variations.add(*product_variation)
+
         else:
             item = CartItem.objects.create(
-                product=product, quantity=1, cart=cart)
+                product=product,
+                quantity=1,
+                user=current_user,
+                cart=cart,
+            )
             if len(product_variation) > 0:
-                # item.variations.clear()
+                item.variations.clear()
                 item.variations.add(*product_variation)
 
+        # item.save()
+        # return redirect('cart')
+
     else:
-        item = CartItem.objects.create(product=product, cart=cart, quantity=1,)
-        if len(product_variation) > 0:
-            # item.variations.clear()
-            item.variations.add(*product_variation)
+        try:
+            cart = Cart.objects.get(cart_id=_cart_id(request))
+        except Cart.DoesNotExist:
+            cart = Cart.objects.create(cart_id=_cart_id(request))
+            cart.save()
+
+        is_given_cart_item_exits = CartItem.objects.filter(
+            product=product, cart=cart).exists()
+
+        ex_var_list = []
+        id = []
+        if is_given_cart_item_exits:
+            cart_item = CartItem.objects.filter(product=product, cart=cart)
+            # existing_variations -> database
+            # current_variation -> product_variation
+            # item_id -> database
+
+            for item in cart_item:
+                existing_variation = item.variations.all()
+                ex_var_list.append(list(existing_variation))
+                id.append(item.id)
+
+            if product_variation in ex_var_list:
+                # increase the cart item quantity
+                index = ex_var_list.index(product_variation)
+                item_id = id[index]
+                item = CartItem.objects.get(product=product, id=item_id)
+                item.quantity += 1
+                # item.save()
+            else:
+                item = CartItem.objects.create(
+                    product=product, quantity=1, cart=cart)
+                if len(product_variation) > 0:
+                    # item.variations.clear()
+                    item.variations.add(*product_variation)
+
+        else:
+            item = CartItem.objects.create(
+                product=product, cart=cart, quantity=1,)
+            if len(product_variation) > 0:
+                item.variations.clear()
+                item.variations.add(*product_variation)
 
     item.save()
 
@@ -78,38 +149,69 @@ def add_cart(request, product_id):
 
 ############################ Minus button functionality of cart #############################
 
-def remove_cart(request, product_id, cart_item_id):
-    cart = Cart.objects.get(cart_id=_cart_id(request))
+def remove_cart(request, cart_item_id):
 
-    # this will return product-id
-    product = get_object_or_404(Product, id=product_id)
+    if request.user.is_authenticated:
+        current_user = request.user
+        cart = CartItem.objects.filter(user=current_user)[0].cart
+        cart_item = CartItem.objects.get(user=current_user, id=cart_item_id)
+        if cart_item.quantity > 1:
+            cart_item.quantity -= 1
+            cart_item.save()
+        else:
+            cart_item.delete()
 
-    try:
-        cart_item = CartItem.objects.get(
-            product=product, cart=cart, id=cart_item_id)
+        is_cartItem_exits_for_given_user = CartItem.objects.filter(
+            user=current_user).exists()
+
+        if is_cartItem_exits_for_given_user == False:
+            cart.delete()
+
+    else:
+
+        cart = Cart.objects.get(cart_id=_cart_id(request))
+
+        cart_item = CartItem.objects.get(id=cart_item_id)
 
         if cart_item.quantity > 1:
             cart_item.quantity -= 1
             cart_item.save()
         else:
             cart_item.delete()
-    except:
-        pass
+
+        is_cartItem_exits = CartItem.objects.filter(cart=cart).exists()
+
+        if is_cartItem_exits == False:
+            cart.delete()
+
     return redirect('cart')
 
 
-# ############################ remove button funtionality of cart########################################
+# ############################ remove button funtionality of cart ########################################
 
-def remove_cart_items(request, product_id, cart_item_id):
-    
-    cart = Cart.objects.get(cart_id=_cart_id(request))
+def remove_cart_items(request, cart_item_id):
 
-    product = get_object_or_404(Product, id=product_id)
+    if request.user.is_authenticated:
+        current_user = request.user
+        cart = CartItem.objects.filter(user=current_user)[0].cart
+        cart_item = CartItem.objects.get(user=current_user, id=cart_item_id)
+        cart_item.delete()
+        is_cartItem_exits_for_given_user = CartItem.objects.filter(
+            user=current_user).exists()
 
-    cart_item = CartItem.objects.get(
-        product=product, cart=cart, id=cart_item_id)
-    
-    cart_item.delete()
+        if is_cartItem_exits_for_given_user == False:
+            cart.delete()
+
+    else:
+        cart = Cart.objects.get(cart_id=_cart_id(request))
+        cart_item = CartItem.objects.get(cart=cart, id=cart_item_id)
+
+        cart_item.delete()
+        is_cartItem_exits = CartItem.objects.filter(cart=cart).exists()
+
+        if is_cartItem_exits == False:
+            cart.delete()
+
     return redirect('cart')
 
 
@@ -118,9 +220,14 @@ def remove_cart_items(request, product_id, cart_item_id):
 def cart(request, total=0, tax=0, grand_total=0, quantity=0, cart_items=None):
 
     try:
-        cart = Cart.objects.get(cart_id=_cart_id(request))
+        if request.user.is_authenticated:
+            cart_items = CartItem.objects.filter(
+                user=request.user, is_active=True)
 
-        cart_items = CartItem.objects.filter(cart=cart, is_active=True)
+        else:
+            cart = Cart.objects.get(cart_id=_cart_id(request))
+            cart_items = CartItem.objects.filter(cart=cart, is_active=True)
+
         for cart_item in cart_items:
             total += (cart_item.product.price * cart_item.quantity)
             quantity += cart_item.quantity
@@ -136,3 +243,31 @@ def cart(request, total=0, tax=0, grand_total=0, quantity=0, cart_items=None):
         'grand_total': grand_total,
     }
     return render(request, 'store/cart.html', context)
+
+
+#################################### Checkout button functionality #############################
+
+@login_required(login_url='login')
+def checkout(request, total=0, quantity=0, cart_items=None):
+
+    try:
+        tax = 0
+        grand_total = 0
+        if request.user.is_authenticated:
+            cart_items = CartItem.objects.filter(
+                user=request.user, is_active=True)
+        for cart_item in cart_items:
+            total += (cart_item.product.price * cart_item.quantity)
+            quantity += cart_item.quantity
+        tax = (2*total)/100
+        grand_total = total + tax
+    except ObjectDoesNotExist:
+        pass
+    context = {
+        'total': total,
+        'quantity': quantity,
+        'cart_items': cart_items,
+        'tax': tax,
+        'grand_total': grand_total,
+    }
+    return render(request, 'store/checkout.html', context)
